--- todo-app/jest.config.ts
@@ -1,7 +1,0 @@
-module.exports = {
-  preset: "ts-jest",
-  testEnvironment: "node",
-  transform: {
-    "^.+\\\\\\\\.tsx?$": "ts-jest",
-  },
-};
